Use custom path for zsh history file
I just realised I lost a tonne of my zsh history a few days ago, and I am angry. I rely on zsh history a lot. The timestamp on the first 987 lines is now 1764292705 (2025-11-28 12:18:25 +1100). That lines up with when I was fiddling with adjustments to the switch-env function in my dotfiles. I think there was a moment where there was a syntax error in it, and it must've not loaded the rest of my dotfiles (including my settings to have unlimited zsh history), and so used the limited zsh history defaults; so then when I ran a command, it bloody truncated my history file. To prevent this from happening again, I'm changing to a custom history file location, so any default config won't apply to it.

https://unix.stackexchange.com/questions/568907/why-do-i-lose-my-zsh-history